Renamed transaction-request folder => transaction folder
