+++ src/transparent-qgw-database/transaction/payer.ts
